refactor: type the SST entry operation instead of carrying a bare byte
SstEntry.op_type and the fourth field of KeyState::Value were a u8 documented
as "0=Put, 1=Insert, 2=Delete". That number was then decoded in two recovery
paths with WalOpKind::from_wire_format, where the same digits name a different
enum: 3 is DeleteRange there. The two numberings agreed only for 0..=2, by
coincidence, and the SST decoder used to accept a Merge byte 3 that the WAL side
would have read as a range delete. (The decoder now refuses it.)

Both fields are now EntryType. The two coverage checks ask what they actually
mean - is this entry a value write? - through EntryType::is_value_write, which is
true for Put and Insert only, so neither path converts through wire bytes and
sst gains no dependency on wal. The memtable adapter maps its own OpType to an
EntryType explicitly instead of reusing the encoded byte, and the reader no
longer casts entry_type to u8.

Forty-one sites in tests and benches carried the literals 0, 1 and 2 and now
name their type; the mapping is the documented one, so no behaviour changes.
Two new tests fail to compile before the change: one pins is_value_write for
all four entry types, and one builds an entry and a key state with a typed
operation.

The 1,708 affected lib tests, the durability, lifecycle and governance suites,
clippy pedantic and the 173-test failpoint fault-injection suite pass.

Not done: the writer API, DynSstWriter::add_with_meta, still takes op_type as a
u8 wire byte, and the callers pass WalOpKind::to_wire_format().

+++ src/sst/fs/reader_io/state.rs
@@ -47,7 +47,7 @@ impl SstFileIo {
                 full_key,
                 value_bytes,
                 entry.sequence,
-                entry.entry_type as u8,
+                entry.entry_type,
                 entry.expiration,
             ));
             offset = next_offset;
@@ -80,7 +80,7 @@ impl SstFileIo {
             block_data.slice(value_offset..value_end),
             entry.sequence,
             entry.expiration,
-            entry.entry_type as u8,
+            entry.entry_type,
         )
     }
